KAN-254 <message: dashboard return api frame>

+++ src/main/java/com/GASB/file/model/dto/response/dashboard/FileDashboardDto.java
@@ -13,14 +13,14 @@ import java.util.List;
 @Builder
 public class FileDashboardDto {
     private int total_count;
-    private float total_volume;
+    private double total_volume;
     private int total_dlp;
     private int total_malware;
     private List<TotalTypeDto> totalTypeDto;
     private List<StatisticsDto> statisticsDto;
 
     @Builder
-    public FileDashboardDto(int total_count, float total_volume, int total_dlp, int total_malware, List<TotalTypeDto> totalTypeDto, List<StatisticsDto> statisticsDto){
+    public FileDashboardDto(int total_count, double total_volume, int total_dlp, int total_malware, List<TotalTypeDto> totalTypeDto, List<StatisticsDto> statisticsDto){
         this.total_count = total_count;
         this.total_volume = total_volume;
         this.total_dlp = total_dlp;
